propagate exceptions thrown during module initialization

+++ litestep/Module.h
@@ -104,7 +104,7 @@ public:
 
 
 private:
-	ULONG CallInit();
+	int CallInit();
 };
 
 #endif // __MODULE_H
